Fix corrupt decoding
On rare occasions, a corrupted packet is received and decoding fails. If
a command doesn't contain the requisite number of parameters or if
conversion of values fails simply ignore the packet (probably ought to
warn in future).

This commit also tweaks the timing of the btcomm loops (mainly to remove
sleep() which is generally not required when events can be waited on
instead).